operator: send CNPNodeStatus updates to k8s
In large scale environments, the current model of having each node in the
cluster try to update the NodeStatus for a CNP means that the amount of requests
sent to update the status via the API server is the following product:

```
\# of CNPs * # of nodes in the cluster
```

This does not scale for large numbers of policies / large numbers of nodes in
the cluster, as this results in a large amount of load on kube-apiserver.

To reduce the amount of interaction directly with the kube-apiserver, have each
ode update the key-value store used by Cilium with its status for the CNP and
have `cilium-operator` be the sole interactor with kube-apiserver for updating
the status. `cilium-operator` will watch for events from the key-value store for
CNP statuses, and will then update the kube-apiserver with the updates.

The information is stored in the key-value store in the following format:

```
cilium/state/cnpstatuses/v2/<UUID of CNP structure>/<namespace>/<name>/<node name> --> <marshaled CNPNodeStatus + namespace, name, UID, node name>
```

The operator has a kvstore watcher which spawns at bootstrap. Upon creation or
modification of a key in the the `cilium/state/cnpstatuses/v2/` prefix, it will
send the update to a goroutine which is responsible for updating the K8s
apiserver with the status of a specific CNP.

The per-CNP updating management is performed by a goroutine which is launched
when the operator receives a creation event for the CNP from its Kubernetes
watcher, and stopped when the watcher receives a deletion event. On deletion,
the prefix of
`cilium/state/cnpstatuses/v2/<UUID of CNP structure>/<namespace>/<name>` is
deleted from the key-value store. If a node is deleted, there is separate logic
already present in `cilium-operator` which will GC old nodes from CNPNodeStatus.

Upon spawning of this goroutine, a shared-store which is launched at operator
bootstrap is used to ensure that we do not miss any events which were pushed for
CNPs which may have received the CNP creation event and pushed them to the
key-value store before the operator received the CNP creation event from its own
watcher.

The logic within the goroutine per-CNP consumes all status events transmitted by
the key-value store watcher until a configurable timeout
(`cnp-status-update-interval`, default 100ms) is hit. It pushes all of these
updates to kube-apiserver in the most efficient / performant way depending on
which Kubernetes version is running. For >=1.13, JSON Patch(es) are used.

Using the cilium-operator to perform these updates allows for multiple status
updates to be sent to the kube-apiserver at once for multiple nodes as opposed
to each node itself updating the kube-apiserver, improving scalability of
CNPNodeStatus updates.

Signed-off by: Ian Vernon <ian@cilium.io>